ExtractRelationshipsFunction/requirements.txt
azure-functions
azure-storage-blob
azure-cosmos
python-dotenv 